ms-schedule/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Excel export
openpyxl==3.1.2

# PDF export
reportlab==4.0.7

# iCal export
icalendar==5.0.11
pytz==2023.3

# Utilities
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Metrics
prometheus-client==0.19.0

# Type hints
typing-extensions==4.9.0

